+++ src/components/domain/Directory/DirectoryDeleteModal/index.tsx
