+++ src/components/domain/Directory/DirectoryListItem/index.tsx
