+++ src/components/domain/Directory/DirectorySidebarListItem/DirectorySidebarListItem.tsx
