+++ src/components/domain/Directory/DirectorySidebarListItem/index.tsx
